+++ src/DevComputaria.Prompts/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace DevComputaria.Prompts;
-
-public class Class1
-{
-
-}